Make explicit casts from (uintptr_t/size_t/time_t) to uint, when we don't care for the truncation or are sure of the results.

+++ trunk/lib/warmux/tools/random.cpp
@@ -45,7 +45,7 @@ void RandomGenerator::InitRandom()
     MSG_DEBUG("random.set", "no, unrandom set");
     return;
   }
-  SetRand(time(NULL));
+  SetRand(uint(time(NULL)));
 }
 
 void RandomGenerator::SetRand(uint seed)
